fix!: update pubsub subscribe method to return subscription
Subscribing to pubsub topics is a two step process - first we subscribe
then we recieve messages.

We may subscribe and not receive messages, so change the return type
to reflect that.

It now returns a promise, after which is resolved the node is subscribed,
and the response has a `messages` method that can be used to iterate over
any received messages, and a `cancel` method that will cancel the
subscription.